Added Homework17Test
