--- src/test/java/Homework18.java
@@ -1,12 +1,0 @@
-public class Homework18 extends BaseTest{
-    @Test
-    public void playASongTest() throws InterruptedException {
-        launchChrome();
-        goToPage();
-        inputEmail("caiman.cotton@testpro.io");
-        inputPassword("te$t$tudent");
-        clickSubmit();
-        playSong();
-
-    }
-}
